refactor(graphql): Add RangeCriteria to simply criteria for ranges
